Code generator figures out types as it goes.
Before this there were two separate parallel functions that figured out the
type and generated the code. This change should make the code generator easier
to maintain.

In this step, the code generator was also simplified. In certain cases it
generates overly complex C-code. But fixing this is an optimization that is not
needed at the moment.