Problem: CCodeGen is confusing.
Before it was only a library for generating C code. Then we
started using it for Dot as well, but it still contained
specific functions for C code.

Solution: Rename it to BlockDoc (it is a library for generating
documents with nested blocks) and move C specifics to C.hs.